requirements.txt
asttokens==2.4.1
black==23.10.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
ipython==8.19.0
jedi==0.19.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyink==23.10.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==2.2.2
